+++ src/mmg2d/mmg2d.h
@@ -199,7 +199,7 @@ int _MMG2D_newElt(MMG5_pMesh mesh);
 int  _MMG2D_delElt(MMG5_pMesh mesh,int iel);
 int _MMG5_getnElt(MMG5_pMesh mesh,int n);
 int MMG2D_zaldy(MMG5_pMesh mesh);
-long long _MMG5_memSize(void);
+size_t _MMG5_memSize(void);
 int _MMG2D_memOption(MMG5_pMesh mesh);
 int  MMG2D_setMeshSize_alloc(MMG5_pMesh);
